+++ src/core/leoImport.ts
@@ -13,6 +13,7 @@ import { command } from './decorators';
 import { Commands } from './leoCommands';
 import { Position } from './leoNodes';
 import { Bead } from './leoUndo';
+import { Block } from '../importers/base_importer';
 //@-<< leoImport imports >>
 //@+others
 //@+node:felix.20230519221548.1: ** class FreeMindImporter
@@ -243,7 +244,7 @@ export class LeoImportCommands {
     //@+node:felix.20230511002352.3: *3* ic.Export
     //@+node:felix.20230511002352.4: *4* ic.convertCodePartToWeb & helpers
     /**
-     * # Headlines not containing a section reference are ignored in noweb
+     * Headlines not containing a section reference are ignored in noweb
      * and generate index index in cweb.
      */
     public convertCodePartToWeb(
@@ -399,7 +400,7 @@ export class LeoImportCommands {
         }
         return [i, result];
     }
-    //@+node:felix.20230511002352.10: *4* ic.convertVnodeToWeb
+    //@+node:felix.20230511002352.10: *4* ic.positionToWeb
     /**
      * This code converts a VNode to noweb text as follows:
 
@@ -439,7 +440,7 @@ export class LeoImportCommands {
                     result += docstart;
                 }
                 [i, result] = this.convertCodePartToWeb(s, i, p, result);
-            } else if (this.treeType === '@file' || startInCode) {
+            } else if (startInCode) {
                 if (!docSeen) {
                     docSeen = true;
                     result += docstart;
@@ -595,18 +596,6 @@ export class LeoImportCommands {
             // theFile = open(fileName, 'w')
             const w_uri = g.makeVscodeUri(fileName);
             let theFile: string = '';
-
-            this.treeType = '@file';
-
-            // Set self.treeType to @root if p or an ancestor is an @root node.
-            for (const p of current.parents()) {
-                let [flag, junk] = g.is_special(p.b, '@root');
-                if (flag) {
-                    this.treeType = '@root';
-                    break;
-                }
-            }
-
             for (const p of current.self_and_subtree(false)) {
                 const s = this.positionToWeb(p);
                 if (s) {
@@ -817,11 +806,12 @@ export class LeoImportCommands {
     public async createOutline(
         parent: Position,
         ext?: string,
-        s?: string
+        s?: string,
+        treeType = '@file',
     ): Promise<Position | undefined> {
         const c = this.c;
         const p = parent.copy();
-        this.treeType = '@file'; // Fix #352.
+        this.treeType = treeType;
         const fileName = c.fullPath(parent);
         const w_isBinary = await g.is_binary_external_file(fileName);
         if (w_isBinary) {
@@ -835,15 +825,15 @@ export class LeoImportCommands {
             return undefined;
         }
 
-        // Each importer file defines `do_import` at the top level with this signature:
-        // def do_import(c: Cmdr, parent: Position, s: str) -> None:
+        // Each importer file defines `do_import` at the top level
 
-        const func = this.dispatch(ext, p); // The do_import callback.
-        // Call the scanning function.
+        const func = this.dispatch(ext, p);
         if (g.unitTesting) {
             // g.assert (func or ext in ('.txt', '.w', '.xxx'), (repr(func), ext, p.h));
             g.assert(func || ['.txt', '.w', '.xxx'].includes(ext), `${func?.toString()}, ${ext}, ${p.h}`);
         }
+
+        // Call the scanning function.
         if (func && !c.config.getBool('suppress-import-parsing', false)) {
             s = g.toUnicode(s, this.encoding);
             s = s.replace(/\r/g, '');
@@ -859,8 +849,7 @@ export class LeoImportCommands {
             return p;
         }
 
-        // #488894: unsettling dialog when saving Leo file
-        // #889175: Remember the full fileName.
+        // Remember the full fileName.
         c.atFileCommands.rememberReadPath(fileName, p);
         p.contract();
         const w = c.frame.body.wrapper;
@@ -1077,7 +1066,7 @@ export class LeoImportCommands {
         files: string[] | undefined,
         parent: Position | undefined,
         shortFn = false,
-        treeType: string | undefined,
+        treeType = '@file',
         verbose = true // Legacy value.
     ): Promise<void> {
         // Not a command.  It must *not* have an event arg.
@@ -1086,7 +1075,7 @@ export class LeoImportCommands {
         if (!c || !c.p || !c.p.__bool__() || !files || !files.length) {
             return;
         }
-        this.treeType = treeType || '@file';
+        this.treeType = treeType;
         this.verbose = verbose;
         if (!parent || !parent.__bool__()) {
             g.trace('===== no parent', g.callers());
@@ -1104,7 +1093,7 @@ export class LeoImportCommands {
                 fn = g.relativeDirectory(c, fn);
                 p.h = `${treeType} ${fn}`;
                 u.afterInsertNode(p, 'Import', undoData);
-                p = await this.createOutline(p);
+                p = await this.createOutline(p, undefined, undefined, treeType);
                 if (p && p.__bool__()) {
                     // createOutline may fail.
                     p.contract();
@@ -1513,55 +1502,141 @@ export class LeoImportCommands {
         }
         return result;
     }
-    //@+node:felix.20230511002352.49: *3* ic.parse_body
+    //@+node:felix.20230511002352.49: *3* ic.parse_body & helpers
     /**
-     * Parse p.b as source code, creating a tree of descendant nodes.
-     * This is essentially an import of p.b.
+     * Split p.b into functions, methods or classes in sibling nodes.
      */
     public parse_body(p: Position): void {
 
         const c = this.c;
-        const d = g.app.language_extension_dict;
         const [u, undoType] = [c.undoer, 'parse-body'];
-        if (!p || !p.__bool__()) {
-            return;
-        }
+        const language = c.getLanguage(p);
 
         if (p.hasChildren()) {
             g.es_print('can not run parse-body: node has children:', p.h);
             return;
         }
-        const language = c.getLanguage(p);
-        this.treeType = '@file';
-        const ext = '.' + d[language];
-        const parser = g.app.classDispatchDict[ext];
-        // Fix bug 151: parse-body creates "None declarations"
-        if (p.isAnyAtFileNode()) {
-            const fn = p.anyAtFileNodeName();
-            [this.methodName, this.fileType] = g.os_path_splitext(fn);
-        } else {
-            const fileType = d[language] || 'py';
-            [this.methodName, this.fileType] = [p.h, fileType];
-        }
-        if (!parser) {
-            g.es_print(
-                `parse-body: no parser for @language ${language || 'None'}`
-            );
+        if (['html', 'xml'].includes(language)) {
+            g.es_print(`parse-body does not support @language ${language}`);
             return;
         }
-        const s = p.b;
+
+        // Instantiate the importer.
+        const importer_class = g.app.importerClassesDict[language];
+        if (!importer_class) {
+            g.es_print(`No importer class for @language ${language}`);
+            return;
+        }
+        const importer = new importer_class(c);
+
         try {
-            const bunch = c.undoer.beforeParseBody(p);
-            p.b = '';
-            parser(c, p, s);
-            u.afterParseBody(p, undoType, bunch);
-            p.expand();
+            const old_p = p.copy();
             c.selectPosition(p);
-            c.redraw();
+            const changed = this.parse_body_helper(p, importer);
+            c.selectPosition(old_p);
+            if (c.checkOutline()) {
+                return; // Error!
+            }
+            if (changed) {
+                u.afterChangeGroup(p, undoType);
+                c.setChanged();
+                c.redraw(old_p);
+            }
         } catch (exception) {
             g.es_exception(exception);
-            p.b = s;
         }
+    }
+    //@+node:felix.20250812231137.1: *4* ic.compute_imported_headline
+    /**
+     * Compute the headline for the given imported lines.
+     */
+    public compute_imported_headline(importer: any, lines: string[], p: Position): string {
+        for (let s of lines) {
+
+            s = s.trim();
+            for (const [kind, pattern] of importer.block_patterns) {
+                let m;
+                if ((m = s.match(pattern))) {
+                    s = m[0];
+                    // Truncate at the first '(' or '{'.
+                    const i1 = s.indexOf('(');
+                    const i2 = s.indexOf('{');
+                    const i = (i1 > -1 && i2 > -1) ? Math.min(i1, i2) : Math.max(i1, i2);
+                    if (i > -1) {
+                        s = s.substring(0, i);
+                    }
+                }
+            }
+        }
+        return p.h;
+    }
+    //@+node:felix.20250812231142.1: *4* ic.parse_body_helper
+    /**
+     * The common code for the parse-body command.
+     */
+    public parse_body_helper(p: Position, importer: any): boolean {
+        // The common code for the parse-body command.
+        const c = this.c;
+        const u = c.undoer;
+        const undoType = 'parse-body';
+
+        importer.lines = g.splitLines(p.b);
+        const lines = importer.lines;
+
+        importer.guide_lines = importer.delete_comments_and_strings(lines);
+        let blocks = importer.find_blocks(0, lines.length);
+
+        // The main loop.
+        let changed = blocks.length > 1;
+        if (changed) {
+            u.beforeChangeGroup(p, undoType);
+        }
+        this.preprocess_blocks(blocks);
+        while (blocks.length > 1) {
+            // Change the node.
+            let bunch = u.beforeChangeBody(p);
+            const block = blocks.shift()!; // pop(0) equivalent
+            const head = lines.slice(block.start, block.end);
+            const tail = lines.slice(block.end);
+
+            p.b = head.join('');
+            p.h = this.compute_imported_headline(importer, head, p);
+            u.afterChangeBody(p, undoType, bunch);
+
+            // Insert another node.
+            bunch = u.beforeInsertNode(p);
+            const p2 = p.insertAfter();
+            p2.h = this.compute_imported_headline(importer, tail, p);
+            p2.b = tail.join('');
+            u.afterInsertNode(p2, undoType, bunch);
+
+            // Continue splitting p2.
+            p = p2;
+        }
+        return changed;
+    }
+    //@+node:felix.20250812231146.1: *4* ic.preprocess_blocks
+    /**
+     * Move blank lines from the start one block to the end of the previous block.
+     */
+    public preprocess_blocks(blocks: Block[]): void {
+
+        for (let i = 0; i < blocks.length; i++) {
+            const block = blocks[i];
+            const block2 = blocks[i + 1];
+            if (!block2) {
+                break; // No next block, mission complete.
+            }
+            for (let j = block2.start; j < block2.end; j++) {
+                const s = block2.lines[j];
+                if (s.trim()) {
+                    break; // Found non-blank line, stop transfer.
+                }
+                block.end += 1;
+                block2.start += 1;
+            }
+        }
+
     }
     //@+node:felix.20230511002352.50: *3* ic.Utilities
     //@+node:felix.20230511002352.51: *4* ic.appendStringToBody & setBodyString (leoImport)
@@ -2211,7 +2286,7 @@ export class RecursiveImportController {
         this.c = c;
         // this.add_path = add_path;
         this.ignore_pattern = ignore_pattern || /\.git|node_modules/;
-        this.kind = kind; // in ('@auto', '@clean', '@edit', '@file', '@nosent')
+        this.kind = kind; // ric.run checks the kind.
         this.n_files = 0;
         this.recursive = recursive;
         this.root = undefined;
@@ -2307,15 +2382,8 @@ export class RecursiveImportController {
         parent: Position
     ): Promise<void> {
         await this.isReady;
-
-        console.log(`Importing file: ${p_path}`);
-
         const c = this.c;
         this.n_files += 1;
-        g.assert(
-            this.root && parent && parent.v !== this.root.v,
-            'Error in import_one_file'
-        );
         let p: Position;
         if (this.kind === '@edit') {
             p = parent.insertAsLastChild();
@@ -2333,11 +2401,18 @@ export class RecursiveImportController {
             [p_path],
             parent,
             true,
-            '@file', // '@auto','@clean','@nosent' cause problems.
+            this.kind,  // Leo 6.8.6.
             this.verbose // Leo 6.6.
         );
+
+        // #4385: set mod time for @clean files. Clear the mod time for all other files.
         p = parent.lastChild();
-        p.h = this.kind + p.h.substring(5); // Honor the requested kind.
+        if (this.kind === '@clean') {
+            p.v.u['_mod_time'] = await g.os_path_getmtime(p_path);
+        } else if ('_mod_time' in p.v.u) {
+            delete p.v.u['_mod_time'];
+        }
+
         if (this.safe_at_file) {
             p.v.h = '@' + p.v.h;
         }
@@ -2676,6 +2751,7 @@ export class RecursiveImportController {
             for (const p2 of parent.self_and_subtree(false)) {
                 p2.contract();
             }
+            c.setChanged();  // #4385: Ensure that mod times are written.
             c.redraw(parent);
         }
         if (!g.unitTesting) {
@@ -2689,68 +2765,6 @@ export class RecursiveImportController {
         }
     }
 
-
-    // def run(self, dir_: Optional[str]) -> None:
-    //     """
-    //     dir_ can be None, a directory contained in the outline's directory, or a single file.
-
-    //     Import all files in the dir_ directory, or the outline's directory if dir_ is None.
-
-    //     Import all files whose extension matches self.theTypes in dir_.
-    //     In fact, dir_ can be a path to a single file.
-    //     """
-    //     if self.kind not in ('@auto', '@clean', '@edit', '@file', '@nosent'):
-    //         self.error(f"bad kind param: {self.kind!r}")
-    //         return
-    //     if not self.outline_directory:
-    //         self.error('The outline has no name. Please save it.')
-    //         return
-
-    //     # Resolve dir_ to an absolute path.
-    //     dir_1 = dir_
-    //     dir_ = self.resolve_dir_arg(dir_)
-    //     if dir_ is None:
-    //         self.error(f"invalid 'dir_' argument: {dir_1!r}")
-    //         return
-
-    //     # Import all requested files.
-    //     try:
-    //         c, u = self.c, self.c.undoer
-    //         t1 = time.time()
-    //         g.app.disable_redraw = True
-    //         last = c.lastTopLevel()
-    //         c.selectPosition(last)
-    //         undoData = u.beforeInsertNode(last)
-    //         # Always create a new last top-level node.
-    //         self.root = parent = last.insertAfter()
-    //         parent.v.h = 'imported files'
-    //         # Special case for a single file.
-    //         self.n_files = 0
-    //         if g.os_path_isfile(dir_):
-    //             if self.verbose:
-    //                 # Only print this message if importing a *single* file.
-    //                 print('')
-    //                 g.es_print(f"importing file: {os.path.normpath(dir_)}")
-    //             self.import_one_file(dir_, parent)
-    //         else:
-    //             self.import_dir(dir_, parent)
-    //         self.post_process(parent)
-    //         u.afterInsertNode(parent, 'recursive-import', undoData)
-    //     except e:
-    //         self.error('Exception in recursive import')
-    //         g.es_exception(e)
-    //     finally:
-    //         g.app.disable_redraw = False
-    //         for p2 in parent.self_and_subtree(copy=False):
-    //             p2.contract()
-    //         c.redraw(parent)
-    //     if not g.unitTesting:
-    //         t2 = time.time()
-    //         n = len(list(parent.subtree()))
-    //         g.es_print(
-    //             f"imported {n} node{g.plural(n)} "
-    //             f"in {self.n_files} file{g.plural(self.n_files)} "
-    //             f"in {t2 - t1:2.2f} seconds")
     //@-others
 }
 //@+node:felix.20230511002653.1: ** class TabImporter